adds possibility to sync to and from the SCIP directory

+++ sync_with_scip.sh
@@ -2,9 +2,32 @@
 
 # this is a simple script to sync all listed Benders' decomposition files with
 # the BendersSCIP repository.
+#
+# By default, the files will be copied from the input SCIP directory to the
+# BendersSCIP directory. To copy the files from the BendersSCIP directory to the
+# SCIP directory, then the -u parameter must be set
 
-SCIPDIR=${1}
-BENDERSFILES=${2}
+UPDATE=0
+SCIPDIR=""
+BENDERSFILES=""
+
+TOORFROM="from"
+
+# parsing the command line arguments
+for i in $@
+do
+   if test "$i" = "-u"
+   then
+      UPDATE=1
+   else
+      if test -z $SCIPDIR
+      then
+         SCIPDIR=$i
+      else
+         BENDERSFILES=$i
+      fi
+   fi
+done
 
 SCIPPATH=`realpath $SCIPDIR`
 
@@ -14,20 +37,46 @@ then
    exit 1
 fi
 
-echo "Syncing all Benders' decomposition framework source files from $SCIPPATH"
+# Checking whether the user is sure they want to update the SCIP directory
+if [ "$UPDATE" = 1 ]
+then
+   echo "This action will update the SCIP directory with the BendersSCIP code."
+   read -p "Are you sure you want to do this? [y/N]: " yn
+
+   # the default is No
+   case $yn in
+      [Yy]* ) break;;
+      [Nn]* ) exit;;
+      * ) exit;;
+   esac
+
+   TOORFROM="to"
+fi
+
+echo "Syncing all Benders' decomposition framework source files $TOORFROM $SCIPPATH"
 
 for file in `cat $BENDERSFILES`
 do
-   SOURCEFILE=$SCIPDIR/$file
+   if [ "$UPDATE" = 0 ]
+   then
+      SOURCEDIR=$SCIPDIR
+      DESTDIR=./
+   else
+      SOURCEDIR=./
+      DESTDIR=$SCIPDIR
+   fi
+
+   SOURCEFILE=$SOURCEDIR/$file
    RELPATH=$(dirname "$file")
+
    # checking whether the source file exists.
    if [ ! -f $SOURCEFILE ]
    then
-      echo "The source file: $file, does not exist."
+      echo "The source file: $SOURCEFILE, does not exist."
    else
       # if the file exists, then we call rsync to copy the file to the repository
-      echo "Syncing: $file"
-      rsync $SOURCEFILE ./$RELPATH/
+      printf 'Syncing: %-45s -> %s\n' $SOURCEDIR$file $(basename $DESTDIR)/$file
+      rsync $SOURCEFILE $DESTDIR/$RELPATH/
    fi
 done
